Moving toolboxPreview file, start refactor.
